Extracted generateToken method to auth.passwordreset module
