+++ core/server/services/auth/index.js
@@ -15,6 +15,10 @@ module.exports = {
         return require('./setup');
     },
 
+    get passwordreset() {
+        return require('./passwordreset');
+    },
+
     /*
      * TODO: Get rid of these when v0.1 is gone
      */
